# app/media_types.go
// Code generated by goagen v1.3.1, DO NOT EDIT.
//
// API "Authentication API": Application Media Types
//
// Command:
// $ goagen
// --design=github.com/rymccue/golang-auth-microservice/design
// --out=$(GOPATH)/src/github.com/rymccue/golang-auth-microservice
// --version=v1.2.0-dirty

package app

import (
	"github.com/goadesign/goa"
)

// A token (default view)
//
// Identifier: application/vnd.token+json; view=default
type Token struct {
	// A JWT token
	Token *string `form:"token,omitempty" json:"token,omitempty" xml:"token,omitempty"`
}

// A user (default view)
//
// Identifier: application/vnd.user+json; view=default
type User struct {
	// Email of the user
	Email *string `form:"email,omitempty" json:"email,omitempty" xml:"email,omitempty"`
	// First name of the user
	FirstName *string `form:"first_name,omitempty" json:"first_name,omitempty" xml:"first_name,omitempty"`
	// ID of account
	ID *int `form:"id,omitempty" json:"id,omitempty" xml:"id,omitempty"`
	// Last name of the user
	LastName *string `form:"last_name,omitempty" json:"last_name,omitempty" xml:"last_name,omitempty"`
	// Avatar of user
	Password *string `form:"password,omitempty" json:"password,omitempty" xml:"password,omitempty"`
	// Phone number of the user
	Salt *string `form:"salt,omitempty" json:"salt,omitempty" xml:"salt,omitempty"`
}

// Validate validates the User media type instance.
func (mt *User) Validate() (err error) {
	if mt.Email != nil {
		if err2 := goa.ValidateFormat(goa.FormatEmail, *mt.Email); err2 != nil {
			err = goa.MergeErrors(err, goa.InvalidFormatError(`response.email`, *mt.Email, goa.FormatEmail, err2))
		}
	}
	return
}